Sample email on casting page Add SubmitDate to Availability Don't update Availability if unchanged

+++ html/set_available_rest.php
@@ -21,18 +21,34 @@ foreach ($_REQUEST as $key => $value) {
 	//echo "Available: " . $key . " Value: " . $value . "<br>\n";
 	list($show) = sscanf($key, "availability%d");
 	//echo "Performer: " . $_SESSION['user_num'] . " Show: " . $show . " Available: " . $value . "<br>\n";
+	
+	// get current value
+	$sql = "SELECT * FROM Availability WHERE Performer = " . $_SESSION['user_num'] . " AND ShowID = " . $show ;
+	// echo $sql . "<br>\n";
+	$checkquery = mysqli_query($conn, $sql)
+		or die (mysqli_error($conn));
 
-	$sql = "DELETE FROM Availability WHERE Performer = " . $_SESSION['user_num'] . " AND ShowID = " . $show ;
-	if ($conn->query($sql) != TRUE) {
-	  echo "Error: " . $sql . "<br>" . $conn->error;
-	  echo "<meta http-equiv=\"refresh\" content=\"15; user_home.php\"/>";
+	if ($availrow = mysqli_fetch_array($checkquery)) {
+		$indb = true;
+		$status = $availrow['Available'];
+	} else {
+		$indb = false;
+		$status = 0;
 	}
-
-	$sql = "INSERT INTO Availability (Performer, ShowID, Available) VALUES (" . $_SESSION['user_num'] . ", " . $show . ", $value)";
-	//echo "SQL: " . $sql . "\n";
-	if ($conn->query($sql) != TRUE) {
-	  echo "Error: " . $sql . "<br>" . $conn->error;
-	  echo "<meta http-equiv=\"refresh\" content=\"15; user_home.php\"/>";
+	if ( $indb && ($status != $value) ) {
+		$sql = "DELETE FROM Availability WHERE Performer = " . $_SESSION['user_num'] . " AND ShowID = " . $show ;
+		if ($conn->query($sql) != TRUE) {
+		  echo "Error: " . $sql . "<br>" . $conn->error;
+		  echo "<meta http-equiv=\"refresh\" content=\"15; user_home.php\"/>";
+		}
+	}
+	if ( ( $indb == false) || ($status != $value)) {
+		$sql = "INSERT INTO Availability (Performer, ShowID, Available) VALUES (" . $_SESSION['user_num'] . ", " . $show . ", $value)";
+		//echo "SQL: " . $sql . "\n";
+		if ($conn->query($sql) != TRUE) {
+		  echo "Error: " . $sql . "<br>" . $conn->error;
+		  echo "<meta http-equiv=\"refresh\" content=\"15; user_home.php\"/>";
+		}
 	}
 }
